A small -Wall cast.
git-svn-id: svn://scm.gforge.inria.fr/svnroot/sollya/trunk@2000 af37bc7e-59aa-4313-bbaf-7d19c0032f32

+++ printf.c
@@ -343,7 +343,7 @@ static inline int specialSnFprintf(FILE *fd, int useFd, char *str, size_t size, 
 
   if (offset < 0) return offset;
 
-  if (offset > size) {
+  if ((size_t) offset > size) {
     actualSize = 0;
     actualStr = NULL;
   } else {
